# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabActivationReason": {
        "enum": [
          "STARTUP_BASELINE",
          "SCENARIO_COMPLETED",
          "RETURN_BASELINE",
          "RESET"
        ],
        "title": "LabActivationReason",
        "type": "string"
      },
      "LabCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "const": "otsoc.final-evaluation.oil-gas-transfer",
            "title": "Dataset Id",
            "type": "string"
          },
          "dataset_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Dataset Sha256",
            "type": "string"
          },
          "dataset_version": {
            "const": "1.0.0",
            "title": "Dataset Version",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LabScenarioCatalogItem"
            },
            "maxItems": 5,
            "minItems": 5,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "dataset_id",
          "dataset_version",
          "dataset_sha256",
          "items"
        ],
        "title": "LabCatalogResponse",
        "type": "object"
      },
      "LabContextResponse": {
        "additionalProperties": false,
        "properties": {
          "activation_reason": {
            "$ref": "#/components/schemas/LabActivationReason"
          },
          "active_run": {
            "$ref": "#/components/schemas/LabRunResponse"
          },
          "changed_at": {
            "format": "date-time",
            "title": "Changed At",
            "type": "string"
          },
          "changed_by_actor": {
            "title": "Changed By Actor",
            "type": "string"
          },
          "changed_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By User Id"
          },
          "context_version": {
            "minimum": 1.0,
            "title": "Context Version",
            "type": "integer"
          }
        },
        "required": [
          "context_version",
          "activation_reason",
          "changed_at",
          "changed_by_user_id",
          "changed_by_actor",
          "active_run"
        ],
        "title": "LabContextResponse",
        "type": "object"
      },
      "LabNoFieldsRequest": {
        "additionalProperties": false,
        "description": "Optional empty body contract for lab mutations that accept no controls.",
        "properties": {},
        "title": "LabNoFieldsRequest",
        "type": "object"
      },
      "LabRunListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LabRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "total"
        ],
        "title": "LabRunListResponse",
        "type": "object"
      },
      "LabRunResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "configuration_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Hash"
          },
          "configuration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Id"
          },
          "dataset_case_id": {
            "title": "Dataset Case Id",
            "type": "string"
          },
          "dataset_id": {
            "const": "otsoc.final-evaluation.oil-gas-transfer",
            "title": "Dataset Id",
            "type": "string"
          },
          "dataset_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Dataset Sha256",
            "type": "string"
          },
          "dataset_version": {
            "const": "1.0.0",
            "title": "Dataset Version",
            "type": "string"
          },
          "definition_version": {
            "const": "1.0.0",
            "title": "Definition Version",
            "type": "string"
          },
          "evidence_count": {
            "minimum": 0.0,
            "title": "Evidence Count",
            "type": "integer"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "incident_count": {
            "minimum": 0.0,
            "title": "Incident Count",
            "type": "integer"
          },
          "incident_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Incident Ids",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scenario_id": {
            "$ref": "#/components/schemas/LabScenarioId"
          },
          "scenario_title": {
            "title": "Scenario Title",
            "type": "string"
          },
          "simulation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "started_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "started_by_display_name": {
            "title": "Started By Display Name",
            "type": "string"
          },
          "started_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By User Id"
          },
          "status": {
            "$ref": "#/components/schemas/LabRunState"
          },
          "window_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          }
        },
        "required": [
          "run_id",
          "scenario_id",
          "scenario_title",
          "definition_version",
          "dataset_id",
          "dataset_version",
          "dataset_sha256",
          "dataset_case_id",
          "simulation_id",
          "configuration_id",
          "configuration_hash",
          "status",
          "started_by_user_id",
          "started_by",
          "started_by_display_name",
          "started_at",
          "completed_at",
          "window_start",
          "window_end",
          "evidence_count",
          "incident_count",
          "incident_ids",
          "failure_code"
        ],
        "title": "LabRunResponse",
        "type": "object"
      },
      "LabRunStartRequest": {
        "additionalProperties": false,
        "properties": {
          "scenario_id": {
            "$ref": "#/components/schemas/LabScenarioId"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "LabRunStartRequest",
        "type": "object"
      },
      "LabRunState": {
        "enum": [
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "LabRunState",
        "type": "string"
      },
      "LabScenarioCatalogItem": {
        "additionalProperties": false,
        "properties": {
          "dataset_case_id": {
            "title": "Dataset Case Id",
            "type": "string"
          },
          "definition_version": {
            "const": "1.0.0",
            "title": "Definition Version",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "execution_mode": {
            "const": "FROZEN_DETERMINISTIC_PIPELINE",
            "title": "Execution Mode",
            "type": "string"
          },
          "scenario_id": {
            "$ref": "#/components/schemas/LabScenarioId"
          },
          "state": {
            "const": "READY",
            "default": "READY",
            "title": "State",
            "type": "string"
          },
          "synthetic": {
            "const": true,
            "title": "Synthetic",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "title",
          "description",
          "dataset_case_id",
          "definition_version",
          "synthetic",
          "execution_mode"
        ],
        "title": "LabScenarioCatalogItem",
        "type": "object"
      },
      "LabScenarioId": {
        "enum": [
          "BASELINE",
          "S1",
          "S2",
          "S3",
          "S4"
        ],
        "title": "LabScenarioId",
        "type": "string"
      },
      "LabStartResponse": {
        "additionalProperties": false,
        "properties": {
          "active_run": {
            "$ref": "#/components/schemas/LabRunResponse"
          },
          "run": {
            "$ref": "#/components/schemas/LabRunResponse"
          }
        },
        "required": [
          "active_run",
          "run"
        ],
        "title": "LabStartResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OTSOCSessionCookie": {
        "description": "Opaque HttpOnly local session cookie. The configured runtime name may differ.",
        "in": "cookie",
        "name": "otsoc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Repository and contract foundation for the isolated, advisory-only OT-SOC Fusion X academic prototype.",
    "title": "OT-SOC Fusion X",
    "version": "1.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/lab/baseline": {
      "post": {
        "operationId": "post_lab_baseline_api_v1_lab_baseline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LabNoFieldsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Post Lab Baseline",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/catalog": {
      "get": {
        "operationId": "get_lab_catalog_api_v1_lab_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Get Lab Catalog",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/context": {
      "get": {
        "operationId": "get_lab_context_api_v1_lab_context_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabContextResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Get Lab Context",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/reset": {
      "post": {
        "operationId": "post_lab_reset_api_v1_lab_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LabNoFieldsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Post Lab Reset",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/runs": {
      "get": {
        "operationId": "get_lab_runs_api_v1_lab_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "scenario_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LabScenarioId"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LabRunState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Get Lab Runs",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/runs/{run_id}": {
      "get": {
        "operationId": "get_lab_run_api_v1_lab_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Get Lab Run",
        "tags": [
          "scenario-lab"
        ]
      }
    },
    "/api/v1/lab/start": {
      "post": {
        "operationId": "post_lab_start_api_v1_lab_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabRunStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Post Lab Start",
        "tags": [
          "scenario-lab"
        ]
      }
    },
...
  }
}